+++ src/OpenTelemetry.Instrumentation.Owin/Implementation/OwinInstrumentationActivitySource.cs
@@ -2,19 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System.Diagnostics;
-using System.Reflection;
-using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Instrumentation.Owin;
 
 internal static class OwinInstrumentationActivitySource
 {
-    internal static readonly Assembly Assembly = typeof(OwinInstrumentationActivitySource).Assembly;
-    internal static readonly AssemblyName AssemblyName = Assembly.GetName();
-    internal static readonly string ActivitySourceName = AssemblyName.Name;
-    internal static readonly string IncomingRequestActivityName = ActivitySourceName + ".IncomingRequest";
+    internal static readonly Version SemanticConventionsVersion = new(1, 41, 0);
 
-    public static ActivitySource ActivitySource { get; } = new(ActivitySourceName, Assembly.GetPackageVersion());
+    internal static readonly ActivitySource ActivitySource = Trace.ActivitySourceFactory.Create(
+        typeof(OwinInstrumentationActivitySource),
+        SemanticConventionsVersion,
+        name: "OpenTelemetry.Instrumentation.Owin.IncomingRequest");
 
     public static OwinInstrumentationOptions? Options { get; set; }
 }